Fixed: gametimer double to float error
nu gameTime voortaan een Float is krijgen we een error van double naar
float, mogelijk verlies..
een double kan namelijk 14 decimalen hebben een float maar 7.
Momenteel verliezen we dus eventueel 0.0000000XXXXXXX de X'jes

Boeit dat? @JohanBeekers @PimV

+++ Bermuda/Bermuda/GameTimer.cpp
@@ -36,7 +36,7 @@ void GameTimer::updateGameTime(double frameTime)
 {
 
 	this->frameTime = frameTime;
-	this->gameTime += frameTime;
+	this->gameTime += static_cast<float>(frameTime);
 	this->updateDay();
 }
